+++ test/Fixture/Methods/NoParameterWithContainerTypeDeclarationRule/Failure/ClassWithMethodWithParameterWithServiceLocatorInterfaceAsTypeDeclaration.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Methods\NoParameterWithContainerTypeDeclarationRule\Failure;
+
+use Zend\ServiceManager;
+
+final class ClassWithMethodWithParameterWithServiceLocatorInterfaceAsTypeDeclaration
+{
+    public function method(ServiceManager\ServiceLocatorInterface $container): void
+    {
+    }
+}