+++ test/Fixture/Methods/NoParameterWithContainerTypeDeclarationRule/Failure/InterfaceExtendingContainerInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Methods\NoParameterWithContainerTypeDeclarationRule\Failure;
+
+use Psr\Container;
+
+interface InterfaceExtendingContainerInterface extends Container\ContainerInterface
+{
+}