+++ test/Fixture/Methods/NoParameterWithContainerTypeDeclarationRule/Failure/InterfaceWithMethodWithParameterWithServiceLocatorInterfaceAsTypeDeclaration.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Localheinz\PHPStan\Rules\Test\Fixture\Methods\NoParameterWithContainerTypeDeclarationRule\Failure;
+
+use Zend\ServiceManager;
+
+interface InterfaceWithMethodWithParameterWithServiceLocatorInterfaceAsTypeDeclaration
+{
+    public function method(ServiceManager\ServiceLocatorInterface $container): void;
+}