+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`0.11.0...master`](https://github.com/localheinz/phpstan-rules/compare/0.11.0...master).
 
+### Added
+
+* Added `Methods\NoParameterWithContainerTypeDeclarationRule`, which reports an error when a method has a type declaration that corresponds to a known dependency injection container or service locator ([#122](https://github.com/localheinz/phpstan-rules/pull/122)), by [@dmecke](https://github.com/dmecke)
+
 ## [`0.11.0`](https://github.com/localheinz/phpstan-rules/releases/tag/0.10.0)
 
 For a full diff see [`0.10.0...0.11.0`](https://github.com/localheinz/phpstan-rules/compare/0.10.0...0.11.0).